backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A backend API server for the-overlord-of-beggars"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
geopy = "^2.4.1"
aiohttp = "^3.10.10"
asyncache = "^0.3.1"
pydantic-settings = "^2.6.0"
xmltodict = "^0.14.2"

[tool.poetry.group.dev.dependencies]
types-xmltodict = "^0.14.0.20241009"
types-cachetools = "^5.5.0.20240820"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
